Responsabilité unique pour les classes de validation de l'entité BetSaved
